services/agent-runtime/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
psycopg[binary]
openai
httpx

